feat: 컴포넌트 네이밍 변경, 구조 변경 (#93)
