+++ src/pages/reservation/ApplicantInfo/index.css.ts
