--- src/pages/reservation/BottomButton/index.css.ts
@@ -1,10 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const bottomButtonStyle = style({
-  width: '100%',
-  justifyItems: 'center',
-  position: 'fixed',
-  bottom: '0',
-  zIndex: 1000,
-  padding: "2.4rem 2rem"
-});
