+++ src/pages/reservation/ClassInfo/index.css.ts
