+++ src/pages/reservation/ClassInfo/index.tsx
