+++ src/pages/reservation/index.css.ts
@@ -9,7 +9,7 @@ export const headerStyle = style({
   zIndex: 10,
 });
 
-export const agreementContainer = style({
+export const agreementContainerStyle = style({
   width: '100%',
 });
 
@@ -41,4 +41,13 @@ export const totalPriceContainer = style({
   width: '100%',
   padding: '1.8rem 2rem 11.8rem 2rem',
   alignItems: 'center',
+});
+
+export const bottomButtonStyle = style({
+  width: '100%',
+  justifyItems: 'center',
+  position: 'fixed',
+  bottom: '0',
+  zIndex: 1000,
+  padding: "2.4rem 2rem"
 });
