+++ src/pages/reservation/index.tsx
@@ -1,19 +1,28 @@
 import { useState } from 'react';
-import InfoComponent from '@/pages/reservation/InfoComponent';
+import { useNavigate } from 'react-router-dom';
+import ClassInfo from '@/pages/reservation/ClassInfo';
+import BoxButton from '@/components/BoxButton';
 import Divider from '@/components/Divider';
 import Flex from '@/components/Flex';
 import Head from '@/components/Head';
 import Header from '@/components/Header';
 import Text from '@/components/Text';
+import { ROUTES_CONFIG } from '@/routes/routesConfig';
 import { IcCheckcircleGray0524, IcCheckcircleMain0324 } from '@/assets/svg';
 import { MY_RESERVATION_DATA } from '@/mocks/mockMyReservationData';
-import AgreeComponent from './AgreeComponent';
-import BookerComponent from './BookerComponent';
-import BottomButton from './BottomButton';
+import AgreeCheckBox from './AgreeCheckBox';
+import ApplicantInfo from './ApplicantInfo';
 import TopImageComponent from './TopImageComponent';
-import { agreementBox, agreementChecked, agreementContainer, agreementUnchecked, headerStyle, reservationStyle, totalPriceContainer } from './index.css';
-import { useNavigate } from "react-router-dom";
-import { ROUTES_CONFIG } from "@/routes/routesConfig";
+import {
+  agreementBox,
+  agreementChecked,
+  agreementContainerStyle,
+  agreementUnchecked,
+  headerStyle,
+  reservationStyle,
+  totalPriceContainer,
+  bottomButtonStyle,
+} from './index.css';
 
 const Reservation = () => {
   const [isAllChecked, setIsAllChecked] = useState(false);
@@ -33,11 +42,14 @@ const Reservation = () => {
 
     setIsAllChecked(newAgreements.every((isChecked) => isChecked));
   };
+
   const { lessonIndividualPrice } = MY_RESERVATION_DATA;
   const navigate = useNavigate();
+
   const handleButtonClick = () => {
     navigate(ROUTES_CONFIG.home.path);
   };
+
   return (
     <Flex direction="column" width="100%" className={reservationStyle}>
       <div className={headerStyle}>
@@ -59,29 +71,27 @@ const Reservation = () => {
           <Text tag="b4" color="gray9">
             클래스 정보
           </Text>
-          <InfoComponent />
+          <ClassInfo />
         </Flex>
         <Flex direction="column" width="100%" gap="1.6rem">
           <Text tag="b4" color="gray9">
             신청자 정보
           </Text>
-          <BookerComponent />
+          <ApplicantInfo />
         </Flex>
       </Flex>
-      <Divider direction="horizontal" length="100%" thickness="1.1rem" />
 
-      {/* 구분선 아래 퍼블리싱 */}
+      <Divider direction="horizontal" length="100%" thickness="1.1rem" color="gray2" />
 
       <Flex direction="column" width="100%" paddingTop="2.8rem" paddingRight="2rem" paddingLeft="2rem">
         <Flex direction="column" width="100%" gap="2rem">
           <Text tag="b4" color="gray9">
             필수 약관 전체 동의
           </Text>
-          <div className={agreementContainer}>
-          <div
+          <div className={agreementContainerStyle}>
+            <div
               onClick={handleToggleAll}
-              className={`${agreementBox} ${isAllChecked ? agreementChecked : agreementUnchecked}`}
-            >
+              className={`${agreementBox} ${isAllChecked ? agreementChecked : agreementUnchecked}`}>
               {isAllChecked ? <IcCheckcircleMain0324 height={24} /> : <IcCheckcircleGray0524 height={24} />}
               <Head level="h5" tag="h6">
                 전체동의
@@ -89,13 +99,13 @@ const Reservation = () => {
             </div>
 
             <Flex direction="column" width="100%">
-              <AgreeComponent
+              <AgreeCheckBox
                 text="개인정보 제공 동의  (필수)"
                 isChecked={agreements[1]}
                 onToggle={() => handleToggle(1)}
                 link="https://youtube.com"
               />
-              <AgreeComponent
+              <AgreeCheckBox
                 text="취소 및 환불 규칙  (필수)"
                 isChecked={agreements[2]}
                 onToggle={() => handleToggle(2)}
@@ -107,7 +117,7 @@ const Reservation = () => {
             * 예약 서비스 이용을 위한 개인정보 수집 및 제 3자 제공, 취소/환불 규정을 확인하였으며 이에 동의합니다.
           </Text>
         </Flex>
-        <Divider direction="horizontal" length="100%" thickness="0.1rem" />
+        <Divider direction="horizontal" length="100%" thickness="0.1rem" color="gray3" />
       </Flex>
 
       <div className={totalPriceContainer}>
@@ -118,11 +128,12 @@ const Reservation = () => {
           {Number(lessonIndividualPrice).toLocaleString()}원{' '}
         </Head>
       </div>
-      <BottomButton
-        isEnabled={isAllChecked}
-        onClick={handleButtonClick}
-      />
-          </Flex>
+      <Flex width="100%" className={bottomButtonStyle}>
+        <BoxButton variant="primary" isDisabled={!isAllChecked} onClick={isAllChecked ? handleButtonClick : undefined}>
+          신청하기
+        </BoxButton>
+      </Flex>
+    </Flex>
   );
 };
